refactor: flatten internal/transport/http/ to internal/http/
The transport/ directory existed to support multiple transports (gRPC, HTTP)
but only HTTP will ever be used. Remove the unnecessary nesting level.
