+++ cmd/server/main.go
@@ -16,8 +16,8 @@ import (
 	"github.com/ErlanBelekov/dist-job-scheduler/internal/infrastructure/postgres"
 	ctxlog "github.com/ErlanBelekov/dist-job-scheduler/internal/log"
 	"github.com/ErlanBelekov/dist-job-scheduler/internal/metrics"
-	httptransport "github.com/ErlanBelekov/dist-job-scheduler/internal/transport/http"
-	"github.com/ErlanBelekov/dist-job-scheduler/internal/transport/http/handler"
+	httptransport "github.com/ErlanBelekov/dist-job-scheduler/internal/http"
+	"github.com/ErlanBelekov/dist-job-scheduler/internal/http/handler"
 	"github.com/ErlanBelekov/dist-job-scheduler/internal/usecase"
 	"github.com/gin-gonic/gin"
 	"github.com/lmittmann/tint"
