+++ internal/http/middleware/ensure_user.go
